- DetermineConanPathFromEnvironment also validates conan executable
