+++ Conan.VisualStudio/ConanOptionsPage.cs
@@ -1,6 +1,5 @@
 using System;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Windows.Forms;
 using Conan.VisualStudio.Core;
 using Microsoft.VisualStudio.Shell;
@@ -16,37 +15,17 @@ namespace Conan.VisualStudio
         private bool? _conanInstallAutomatically;
         private ConanBuildType? _conanBuild;
         private bool? _conanUpdate;
-        private static bool ValidateConanExecutable(string exe)
+
+
+        private bool ValidateConanExecutableAndShowMessage(string exe)
         {
-            if (exe == null || exe == "")
-                return true;
-            try
+            if (!ConanPathHelper.ValidateConanExecutable(exe, out string errorMessage))
             {
-                var startInfo = new ProcessStartInfo
-                {
-                    FileName = exe,
-                    Arguments = "--version",
-                    UseShellExecute = false,
-                    RedirectStandardInput = true,
-                    RedirectStandardOutput = true,
-                    CreateNoWindow = true
-                };
-                var process = Process.Start(startInfo);
-
-                process.WaitForExit();
-
-                if (0 != process.ExitCode)
-                    MessageBox.Show($"invalid conan executable {exe}: conan --version failed with error {process.ExitCode}",
-                        "invalid conan executable", MessageBoxButtons.OK, MessageBoxIcon.Error);
-
-                return 0 == process.ExitCode;
-            }
-            catch (Win32Exception e)
-            {
-                MessageBox.Show($"invalid conan executable {exe}: {e.Message}", "invalid conan executable",
+                MessageBox.Show(errorMessage, "invalid conan executable",
                     MessageBoxButtons.OK, MessageBoxIcon.Error);
                 return false;
             }
+            return true;
         }
 
         [Category("Conan")]
@@ -55,7 +34,7 @@ namespace Conan.VisualStudio
         public string ConanExecutablePath
         {
             get => _conanExecutablePath ?? (_conanExecutablePath = ConanPathHelper.DetermineConanPathFromEnvironment());
-            set => _conanExecutablePath = ValidateConanExecutable(value) ? value : _conanExecutablePath;
+            set => _conanExecutablePath = ValidateConanExecutableAndShowMessage(value) ? value : _conanExecutablePath;
         }
 
         [Category("Conan")]
